+++ CHANGES.md
@@ -30,7 +30,10 @@
 
 - `--diff` correctly indicates when a file doesn't end in a newline (#1662)
 
+- Added `--stdin-filename` argument to allow stdin to respect `--force-exclude` rules
+
 - Reduce second-pass formatting stability errors (#1629)
+  (#1780)
 
 #### _Packaging_